+++ app/models.py
@@ -283,6 +283,9 @@ class Service(db.Model, Versioned):
         default_letter_contact = [x for x in self.letter_contacts if x.is_default]
         return default_letter_contact[0].contact_block if default_letter_contact else None
 
+    def has_permission(self, permission):
+        return permission in [p.permission for p in self.permissions]
+
 
 class AnnualBilling(db.Model):
     __tablename__ = "annual_billing"
